+++ AnimatedTangle/patterns/doors.js
@@ -3,7 +3,7 @@ import { Point } from "../../sketchlib/pga2d/Point.js";
 import { Ease } from "../../sketchlib/Ease.js";
 import { GroupPrimitive } from "../../sketchlib/primitives/GroupPrimitive.js";
 import { PolygonPrimitive } from "../../sketchlib/primitives/PolygonPrimitive.js";
-import { Rect } from "../../sketchlib/primitives/RectPrimitive.js";
+import { Rect } from "../../sketchlib/primitives/Rect.js";
 import { group, style, xform } from "../../sketchlib/primitives/shorthand.js";
 import { Transform } from "../../sketchlib/primitives/Transform.js";
 import { Style } from "../../sketchlib/Style.js";
