+++ AnimatedTangle/patterns/seascape.js
@@ -15,7 +15,7 @@ import {
 import { mod } from "../../sketchlib/mod.js";
 import { GroupPrimitive } from "../../sketchlib/primitives/GroupPrimitive.js";
 import { BeziergonPrimitive } from "../../sketchlib/primitives/BeziergonPrimitive.js";
-import { Rect } from "../../sketchlib/primitives/RectPrimitive.js";
+import { Rect } from "../../sketchlib/primitives/Rect.js";
 import { Circle } from "../../sketchlib/primitives/Circle.js";
 import { LineSegment } from "../../sketchlib/primitives/LineSegment.js";
 import { Animated } from "../../sketchlib/animation/Animated.js";
