+++ Coral/rendering.js
@@ -2,7 +2,7 @@ import { Rect } from "./Rect.js";
 import { CoralTile } from "./CoralTile.js";
 import { CardinalDirection } from "../sketchlib/CardinalDirection.js";
 import { LineSegment } from "../sketchlib/primitives/LineSegment.js";
-import { Rect } from "../sketchlib/primitives/RectPrimitive.js";
+import { Rect } from "../sketchlib/primitives/Rect.js";
 import { Point } from "../sketchlib/pga2d/Point.js";
 
 /**
