+++ sketchlib/input/ToggleButton.js
@@ -1,7 +1,7 @@
 import { Point } from "../pga2d/Point.js";
 import { Color } from "../Color.js";
 import { GroupPrimitive } from "../primitives/GroupPrimitive.js";
-import { Rect } from "../primitives/RectPrimitive.js";
+import { Rect } from "../primitives/Rect.js";
 import { style } from "../primitives/shorthand.js";
 import { Style } from "../Style.js";
 import { Rectangle } from "../Rectangle.js";
