+++ sketchlib/primitives/Rect.js
@@ -1,5 +1,5 @@
-import { Direction } from "../../sketchlib/pga2d/Direction.js";
-import { Point } from "../../sketchlib/pga2d/Point.js";
+import { Direction } from "../pga2d/Direction.js";
+import { Point } from "../pga2d/Point.js";
 import { HEIGHT, WIDTH } from "../dimensions.js";
 import { clamp } from "../clamp.js";
 import { Grid } from "../Grid.js";
